chore(api): extract test-side encrypt helper to shared module
Pure code-shape refactor. Extract the AES-256-GCM test-side encrypt
helper from 4 duplicate locations into a single shared module at
__tests__/helpers/encrypt-secret.ts. Standardize on the canonical name
encryptSecretForTests across all 4 call sites.

Affected helpers:
- zero-slack-channels.ts (was: encryptBotTokenForTests)
- zero-telegram.ts (was: encryptBotTokenForTests)
- zero-model-providers.ts (was: encryptSecretValueForTests)
- zero-integrations-slack.ts (was: encryptSecretForTests, already canonical)

Each helper now imports from "./encrypt-secret" and drops its local
copy plus the now-unused createCipheriv / randomBytes / env imports.

No behavior change. The 5 affected test files
(zero-slack-channels.test.ts, zero-telegram-data.service.test.ts,
zero-integrations-telegram.test.ts, zero-model-providers.test.ts,
zero-integrations-slack-status.test.ts) pass without modification —
that is the verification discipline for this refactor.

Closes #12403
