+++ turbo/apps/api/src/signals/routes/__tests__/helpers/encrypt-secret.ts
@@ -1,0 +1,19 @@
+import { createCipheriv, randomBytes } from "node:crypto";
+
+import { env } from "../../../../lib/env";
+
+export function encryptSecretForTests(plaintext: string): string {
+  const key = Buffer.from(env("SECRETS_ENCRYPTION_KEY"), "hex");
+  const iv = randomBytes(12);
+  const cipher = createCipheriv("aes-256-gcm", key, iv, { authTagLength: 16 });
+  const data = Buffer.concat([
+    cipher.update(plaintext, "utf8"),
+    cipher.final(),
+  ]);
+  const authTag = cipher.getAuthTag();
+  return [
+    iv.toString("base64"),
+    authTag.toString("base64"),
+    data.toString("base64"),
+  ].join(":");
+}
